% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPPA_datasets.R
\docType{data}
\name{RPPA}
\alias{ACC.RPPA}
\alias{ACC.RPPA,BLCA.RPPA,BRCA.RPPA,CESC.RPPA,CHOL.RPPA,COAD.RPPA,COADREAD.RPPA,DLBC.RPPA,ESCA.RPPA,GBM.RPPA,GBMLGG.RPPA,HNSC.RPPA,KICH.RPPA,KIPAN.RPPA,KIRC.RPPA,KIRP.RPPA,LGG.RPPA,LIHC.RPPA,LUAD.RPPA,LUSC.RPPA,MESO.RPPA,OV.RPPA,PAAD.RPPA,PCPG.RPPA,PRAD.RPPA,READ.RPPA,SARC.RPPA,SKCM.RPPA,STAD.RPPA,STES.RPPA,TGCT.RPPA,THCA.RPPA,THYM.RPPA,UCEC.RPPA,UCS.RPPA,UVM.RPPA}
\alias{BLCA.RPPA}
\alias{BRCA.RPPA}
\alias{CESC.RPPA}
\alias{CHOL.RPPA}
\alias{COAD.RPPA}
\alias{COADREAD.RPPA}
\alias{DLBC.RPPA}
\alias{ESCA.RPPA}
\alias{GBM.RPPA}
\alias{GBMLGG.RPPA}
\alias{HNSC.RPPA}
\alias{KICH.RPPA}
\alias{KIPAN.RPPA}
\alias{KIRC.RPPA}
\alias{KIRP.RPPA}
\alias{LGG.RPPA}
\alias{LIHC.RPPA}
\alias{LUAD.RPPA}
\alias{LUSC.RPPA}
\alias{MESO.RPPA}
\alias{OV.RPPA}
\alias{PAAD.RPPA}
\alias{PCPG.RPPA}
\alias{PRAD.RPPA}
\alias{READ.RPPA}
\alias{RPPA}
\alias{SARC.RPPA}
\alias{SKCM.RPPA}
\alias{STAD.RPPA}
\alias{STES.RPPA}
\alias{TGCT.RPPA}
\alias{THCA.RPPA}
\alias{THYM.RPPA}
\alias{UCEC.RPPA}
\alias{UCS.RPPA}
\alias{UVM.RPPA}
\title{RPPA datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.RPPA

BLCA.RPPA

BRCA.RPPA

CESC.RPPA

CHOL.RPPA

COAD.RPPA

COADREAD.RPPA

DLBC.RPPA

ESCA.RPPA

GBM.RPPA

GBMLGG.RPPA

HNSC.RPPA

KICH.RPPA

KIPAN.RPPA

KIRC.RPPA

KIRP.RPPA

LGG.RPPA

LIHC.RPPA

LUAD.RPPA

LUSC.RPPA

MESO.RPPA

OV.RPPA

PAAD.RPPA

PCPG.RPPA

PRAD.RPPA

READ.RPPA

SARC.RPPA

SKCM.RPPA

STAD.RPPA

STES.RPPA

TGCT.RPPA

THCA.RPPA

THYM.RPPA

UCEC.RPPA

UCS.RPPA

UVM.RPPA
}
\value{
Data frames with RPPA data.
}
\description{
Package provides RPPA datasets from The Cancer Genome Atlas Project for
all available cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots
for the date: \code{2015-11-01}. The process is described here:
\url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.data}
family can be found here: \code{browseVignettes("RTCGA")}. RPPA data format
is explained here
\url{https://wiki.nci.nih.gov/display/TCGA/Protein+Array+Data+Format+Specification?src=search}.
Converting \pkg{RTCGA.RPPA} to Bioconductor datasets' format is explained here \link{convertTCGA}.
}
\examples{

\dontrun{
browseVignettes("RTCGA")
browseVignettes("RTCGA.RPPA")
}

}
\keyword{datasets}

